--- a/Cargo.toml
+++ b/Cargo.toml
@@ -118,6 +118,7 @@ cesu8 = "1.1.0"
 cfb8 = "0.8.1"
 clap = { version = "4.0.30", features = ["derive"] }
 criterion = "0.5.1"
+derive_more = "0.99.17"
 directories = "5.0.0"
 eframe = { version = "0.22.0", default-features = false }
 egui = "0.22.0"