--- a/crates/valence_entity/Cargo.toml
+++ b/crates/valence_entity/Cargo.toml
@@ -13,6 +13,7 @@ anyhow.workspace = true
 bevy_app.workspace = true
 bevy_ecs.workspace = true
 bitfield-struct.workspace = true
+derive_more.workspace = true
 valence_math.workspace = true
 paste.workspace = true
 rustc-hash.workspace = true