--- a/crates/valence_entity/src/hitbox.rs
+++ b/crates/valence_entity/src/hitbox.rs
@@ -2,9 +2,9 @@
 
 use bevy_app::prelude::*;
 use bevy_ecs::prelude::*;
+use derive_more::Deref;
 use valence_math::{Aabb, DVec3, UVec3, Vec3Swizzles};
 use valence_protocol::Direction;
-use derive_more::Deref;
 
 use crate::*;