--- a/crates/valence_inventory/Cargo.toml
+++ b/crates/valence_inventory/Cargo.toml
@@ -11,5 +11,6 @@ license.workspace = true
 [dependencies]
 bevy_app.workspace = true
 bevy_ecs.workspace = true
+derive_more.workspace = true
 tracing.workspace = true
 valence_server.workspace = true