--- a/crates/valence_player_list/Cargo.toml
+++ b/crates/valence_player_list/Cargo.toml
@@ -12,4 +12,5 @@ license.workspace = true
 bevy_app.workspace = true
 bevy_ecs.workspace = true
 bitfield-struct.workspace = true
+derive_more.workspace = true
 valence_server.workspace = true