--- a/crates/valence_server_common/Cargo.toml
+++ b/crates/valence_server_common/Cargo.toml
@@ -11,6 +11,7 @@ license.workspace = true
 [dependencies]
 bevy_ecs.workspace = true
 bevy_app.workspace = true
+derive_more.workspace = true
 uuid.workspace = true
 rand.workspace = true
 valence_protocol.workspace = true