--- a/crates/valence_server_common/src/uuid.rs
+++ b/crates/valence_server_common/src/uuid.rs
@@ -1,4 +1,5 @@
 use bevy_ecs::prelude::*;
+use derive_more::Deref;
 use uuid::Uuid;
 
 /// The universally unique identifier of an entity. Component wrapper for a
@@ -6,7 +7,7 @@ use uuid::Uuid;
 ///
 /// This component is expected to remain _unique_ and _constant_ during the
 /// lifetime of the entity. The [`Default`] impl generates a new random UUID.
-#[derive(Component, Copy, Clone, PartialEq, Eq, Debug, PartialOrd, Ord, Hash)]
+#[derive(Component, Copy, Clone, PartialEq, Eq, Debug, PartialOrd, Ord, Hash, Deref)]
 pub struct UniqueId(pub Uuid);
 
 /// Generates a new random UUID.