--- a/crates/valence_weather/Cargo.toml
+++ b/crates/valence_weather/Cargo.toml
@@ -12,3 +12,4 @@ license.workspace = true
 valence_server.workspace = true
 bevy_ecs.workspace = true
 bevy_app.workspace = true
+derive_more.workspace = true