--- a/crates/valence_world_border/Cargo.toml
+++ b/crates/valence_world_border/Cargo.toml
@@ -11,4 +11,5 @@ license.workspace = true
 [dependencies]
 bevy_app.workspace = true
 bevy_ecs.workspace = true
+derive_more.workspace = true
 valence_server.workspace = true