--- a/crates/valence_world_border/src/lib.rs
+++ b/crates/valence_world_border/src/lib.rs
@@ -28,6 +28,7 @@ use valence_server::protocol::packets::play::{
 };
 use valence_server::protocol::WritePacket;
 use valence_server::{ChunkLayer, Server};
+use derive_more::{Deref, DerefMut};
 
 // https://minecraft.fandom.com/wiki/World_border
 pub const DEFAULT_PORTAL_LIMIT: i32 = 29999984;
@@ -96,7 +97,7 @@ pub struct WorldBorderLerp {
     /// automatically.
     pub remaining_ticks: u64,
 }
-#[derive(Component, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Debug)]
+#[derive(Component, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Debug, Deref, DerefMut)]
 pub struct WorldBorderWarnTime(pub i32);
 
 impl Default for WorldBorderWarnTime {
@@ -105,7 +106,7 @@ impl Default for WorldBorderWarnTime {
     }
 }
 
-#[derive(Component, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Debug)]
+#[derive(Component, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Debug, Deref, DerefMut)]
 pub struct WorldBorderWarnBlocks(pub i32);
 
 impl Default for WorldBorderWarnBlocks {
@@ -114,7 +115,7 @@ impl Default for WorldBorderWarnBlocks {
     }
 }
 
-#[derive(Component, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Debug)]
+#[derive(Component, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Debug, Deref, DerefMut)]
 pub struct WorldBorderPortalTpBoundary(pub i32);
 
 impl Default for WorldBorderPortalTpBoundary {