## Objective

- Add `Deref` and `DerefMut` to all components using `derive_more` crate
- Closes #444

## Solution

- Add `derive_more` crate to top level `Cargo.toml` and use cargo workspace dependency in necessary crates
- Derive `DerefMut` trait to components that provide mutable access to underlying data
- Derive `Deref` trait to components that provide immutable access to underlying data

## Compile times

I made some compilation time benchmarks on my machine (i don't know good way to do compile time benchmarks so i ran debug mode compilation couple of times before and after the change)

Clean `Valence` compilation
- Before: 2m 2s
- After: 2m 5s

Clean default `playground` example
- Before: 11s
- After: 11s