@dyc3 2023-08-18
Good to see that it doesn't really seem to affect compile times. What CPU do you have?