@IoaNNUwU 2023-08-18
My CPU is Ryzen 5500