@rj00a 2023-08-18
I would prefer to use the `1.0.0-beta.3` version of `derive_more` here.