@dyc3 2023-08-18
I just took a look at the [`derive_more` changelog](https://github.com/JelteF/derive_more/blob/v1.0.0-beta.3/CHANGELOG.md). The only thing I saw that could affect us here is that the default for v1 is to have all features disabled, but I think that's a minor issue.