@rj00a 2023-08-19
I would actually suggest using the macros from `bevy_derive` since we already transitively depend on it, but I recently started using `derive_more 1.0.0-beta.3` in my `better-packets` branch for `valence_protocol`.