@rj00a 2023-08-19
Looks like this missed the generated components in `valence_entity`.